+++ src/components/Main/FollowingPage.jsx
@@ -5,10 +5,10 @@ import { getFollowingUsersData } from "../../apis/user";
 import UsersGrid from "./User/UsersGrid";
 import Pagination from "./Common/Pagination";
 import NullUserGrid from "./NullUserGrid";
+import usePagination from "../../hooks/usePagination";
 
 export default function FollowingPage() {
   const [page, setPage] = useState(1);
-  const [pageStart, setPageStart] = useState(1);
   const {
     data: { data: userData, totalPage },
   } = useSuspenseQuery({
@@ -16,41 +16,14 @@ export default function FollowingPage() {
     queryFn: () => getFollowingUsersData({ pageNumber: page }),
     keepPreviousData: true,
   });
-
-  const getItemProps = (value) => ({
-    variant: page === value ? "filled" : "text",
-    color: "white",
-    className: `${page === value ? "bg-brand text-white" : ""}`,
-    ripple: false,
-    onClick: () => {
-      setPage(value);
-      if (value < page) {
-        if (page % 5 === 1) setPageStart(page - 5);
-      } else {
-        if (page % 5 === 0) setPageStart(page + 1);
-      }
-    },
-  });
-  const handleNextClick = () => {
-    if (page === totalPage) return;
-
-    if (page % 5 === 0) setPageStart(page + 1);
-    setPage((prev) => prev + 1);
-  };
-  const handleLastPageClick = () => {
-    setPage(totalPage);
-    setPageStart(totalPage - ((totalPage % 5) - 1));
-  };
-  const handlePrevClick = () => {
-    if (page === 1) return;
-
-    if (page % 5 === 1) setPageStart(page - 5);
-    setPage((prev) => prev - 1);
-  };
-  const handleFirstPageClick = () => {
-    setPage(1);
-    setPageStart(1);
-  };
+  const {
+    pageStart,
+    getItemProps,
+    handleNextClick,
+    handleFirstPageClick,
+    handleLastPageClick,
+    handlePrevClick,
+  } = usePagination({ page, setPage, totalPage });
   return (
     <section className="flex flex-col justify-between gap-4 h-full">
       {userData.length === 0 ? (
